fix movement being contingent on client transmission rate
