+++ src/Client.ts
@@ -256,24 +256,6 @@ export default class Client {
                 this.inputs.mouse.y = util.constrain(mouseY, minY, maxY);
 
                 if (!Vector.isFinite(this.inputs.mouse)) break;
-                const movement: VectorAbstract = {
-                    x: 0,
-                    y: 0
-                };
-
-                if (flags & InputFlags.up) movement.y -= 1;
-                if (flags & InputFlags.down) movement.y += 1;
-                if (flags & InputFlags.right) movement.x += 1;
-                if (flags & InputFlags.left) movement.x -= 1;
-                
-                if (movement.x || movement.y) {
-                    const angle = Math.atan2(movement.y, movement.x);
-
-                    const magnitude = util.constrain(Math.sqrt(movement.x ** 2 + movement.y ** 2), -1, 1);
-
-                    this.inputs.movement.magnitude = magnitude;
-                    this.inputs.movement.angle = angle;
-                }
 
                 const player = camera.cameraData.values.player;
                 if (!Entity.exists(player) || !(player instanceof TankBody)) return;
