+++ src/Entity/Tank/TankBody.ts
@@ -36,6 +36,7 @@ import { Inputs } from "../AI";
 import AbstractBoss from "../Boss/AbstractBoss";
 import { ArenaState } from "../../Native/Arena";
 import { AccessLevel, maxPlayerLevel } from "../../config";
+import { VectorAbstract } from "../../Physics/Vector";
 
 /**
  * Abstract type of entity which barrels can connect to.
@@ -252,6 +253,22 @@ export default class TankBody extends LivingEntity implements BarrelBase {
     }
 
     public tick(tick: number) {
+        const movement: VectorAbstract = { x: 0, y: 0 };
+
+        if (this.inputs.flags & InputFlags.up) movement.y -= 1;
+        if (this.inputs.flags & InputFlags.down) movement.y += 1;
+        if (this.inputs.flags & InputFlags.right) movement.x += 1;
+        if (this.inputs.flags & InputFlags.left) movement.x -= 1;
+        
+        if (movement.x || movement.y) {
+            const angle = Math.atan2(movement.y, movement.x);
+
+            const magnitude = util.constrain(Math.sqrt(movement.x ** 2 + movement.y ** 2), -1, 1);
+
+            this.inputs.movement.magnitude = magnitude;
+            this.inputs.movement.angle = angle;
+        }
+        
         this.positionData.angle = Math.atan2(this.inputs.mouse.y - this.positionData.values.y, this.inputs.mouse.x - this.positionData.values.x);
 
         if (this.isInvulnerable) {
